Extract rpc_utils from prove_block
